Refactor the Interpreter implementation
