+++ pkg/interpreter/interpreter.go
@@ -1,7 +1,8 @@
-package py
+package interpreter
 
 import (
 	"fmt"
+	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -103,10 +104,10 @@ func (i Interpreter) SatisfiesExact(major, minor int) bool {
 // and enables us to implement sorting which is how we tell which one is
 // the latest python version without relying on filesystem lexical order
 // which may not be deterministic
-type InterpreterList []Interpreter
+type List []Interpreter
 
 // Len returns the number of interpreters in the list
-func (il InterpreterList) Len() int {
+func (il List) Len() int {
 	return len(il)
 }
 
@@ -114,7 +115,7 @@ func (il InterpreterList) Len() int {
 // less than element with index j
 // Note: we reverse it here and actually test for greater than
 // because we want the latest interpreter to be at the front of the slice
-func (il InterpreterList) Less(i, j int) bool {
+func (il List) Less(i, j int) bool {
 	// Short circuit, if i.Major > j.Major, return true straight away
 	if il[i].Major > il[j].Major {
 		return true
@@ -132,6 +133,93 @@ func (il InterpreterList) Less(i, j int) bool {
 }
 
 // Swap swaps the position of two elements in the list
-func (il InterpreterList) Swap(i, j int) {
+func (il List) Swap(i, j int) {
 	il[i], il[j] = il[j], il[i]
+}
+
+// GetAll looks under each path in `paths` for valid python
+// interpreters and returns the ones it finds
+//
+// A valid python interpreter in this context is any filepath with a base name
+// that starts with `python`
+// This is allowed in this context because in usage in this program, `paths` will
+// be populated by searching through $PATH, meaning we don't have to bother checking
+// if files are executable etc and $PATH is unlikely to be cluttered with random
+// files called `python` unless they are the interpreter executables
+func GetAll(paths []string) (List, error) {
+	var interpreters List
+
+	for _, path := range paths {
+		found, err := getPythonInterpreters(path)
+		if err != nil {
+			return nil, fmt.Errorf("could not fetch interpreters under %s: %w", path, err)
+		}
+		interpreters = append(interpreters, found...)
+	}
+
+	return interpreters, nil
+}
+
+// getPythonInterpreters accepts an absolute path to a directory under which
+// it will search for python interpreters, returning any it finds
+func getPythonInterpreters(dir string) (List, error) {
+	contents, err := os.ReadDir(dir)
+	if err != nil {
+		return nil, fmt.Errorf("could not read contents of %s: %w", dir, err)
+	}
+
+	var interpreters []Interpreter
+
+	for _, item := range contents {
+		var interpreter Interpreter
+		itemPath := filepath.Join(dir, item.Name())
+		if err := interpreter.FromFilePath(itemPath); err == nil {
+			// Only add if the interpreter is valid, the others we don't care about
+			interpreters = append(interpreters, interpreter)
+		}
+	}
+
+	return interpreters, nil
+}
+
+// deDupe takes in a list of paths (e.g. those returned from GetPath)
+// and returns a de-duplicated list
+// it is not that common to have a duplicated $PATH entry but it could happen
+// so let's handle it here
+func deDupe(paths []string) []string {
+	keys := make(map[string]bool)
+	deDuped := []string{}
+	for _, item := range paths {
+		if _, ok := keys[item]; !ok {
+			keys[item] = true
+			deDuped = append(deDuped, item)
+		}
+	}
+
+	return deDuped
+}
+
+// GetPath looks up the $PATH environment variable and will return
+// each unique path in a string slice
+func GetPath(key string) ([]string, error) {
+	path, ok := os.LookupEnv(key)
+	if !ok {
+		// This should literally never happen on any Unix system
+		return nil, fmt.Errorf("could not get $%s", key)
+	}
+
+	paths := []string{}
+
+	for _, dir := range filepath.SplitList(path) {
+		if dir == "" {
+			// Unix shell semantics: path element "" means "."
+			dir = "."
+		}
+		paths = append(paths, dir)
+	}
+
+	// Dedupe
+	paths = deDupe(paths)
+
+	return paths, nil
 }
